Tweak how roots of call graphs are computed
Currently, a recursive procedure that has no callers is not considered
a root.  This results in CallGraph.GetPostOrder returning fewer
procedures than are present in the call graph.  Due to this, SymDiff
doesn't properly compute mod sets and skips generating equivalence
checks for certain pairs of inputs.

This commit fixes the issue by modifying CallGraphNode.GetChildren to
exclude the current node.  When the current node corresponds to a
recursive procedure, SimpleGraph.FindRoots will not skip it.