Flag to skip unaligned globals for equiv checking
This commit adds the flag `-ignoreUnalignedGlobals`.  When checking
equivalence between procedures `p1` and `p2`, SymDiff checks that all
modified globals that aren't present in the config don't change.

For example, if `p2` modifies a global `$g` that isn't mapped to any
other global in the config, SymDiff will check that `old($g) == $g`.
The flag `-ignoreUnalignedGlobals` disables this behavior, causing
SymDiff to not check any condition on `$g`.